feat: produce result with alert, non_detections and prv_detections
