+++ prv_candidates_step/core/strategy/atlas_strategy.py
@@ -1,11 +1,6 @@
 from .base_strategy import BasePrvCandidatesStrategy
 
-import pandas as pd
-
-# Keys used on non detections for ZTF
-NON_DET_KEYS = ["aid", "tid", "oid", "mjd", "diffmaglim", "fid"]
-
 
 class ATLASPrvCandidatesStrategy(BasePrvCandidatesStrategy):
-    def process_prv_candidates(self, alerts: pd.DataFrame):
-        return pd.DataFrame(), pd.DataFrame(columns=NON_DET_KEYS)
+    def process_prv_candidates(self, _: dict):
+        return [], []
