+++ prv_candidates_step/step.py
@@ -1,13 +1,13 @@
+from typing import Tuple
 from apf.core.step import GenericStep
 from prv_candidates_step.core.candidates.process_prv_candidates import (
     process_prv_candidates,
 )
 from prv_candidates_step.core.strategy.ztf_strategy import ZTFPrvCandidatesStrategy
 from prv_candidates_step.core.processor.processor import Processor
+from prv_candidates_step.core import ZTFPreviousDetectionsParser, ZTFNonDetectionsParser
 
 
-import numpy as np
-import pandas as pd
 import logging
 
 
@@ -35,21 +35,32 @@ class PrvCandidatesStep(GenericStep):
         )  # initial strategy (can change)
         self.producers = {"scribe": None, "alerts": None}
 
-    def pre_produce(self, result: pd.DataFrame):
+    def pre_produce(self, result: Tuple):
         self.set_producer_key_field("aid")
-        return result.to_dict("records")
+        output = []
+        for index, alert in enumerate(result[0]):
+            ztf_prv_detections_parser = ZTFPreviousDetectionsParser()
+            ztf_non_detections_parser = ZTFNonDetectionsParser()
+            parsed_prv_detections = ztf_prv_detections_parser.parse(
+                result[1][index], alert["oid"], alert["aid"], alert["candid"]
+            )
+            parsed_non_detections = ztf_non_detections_parser.parse(
+                result[2][index], alert["aid"], alert["tid"], alert["oid"]
+            )
+            output.append(
+                {
+                    "aid": alert["aid"],
+                    "new_alert": alert,
+                    "prv_detections": parsed_prv_detections,
+                    "non_detections": parsed_non_detections,
+                }
+            )
+        return output
 
     def execute(self, messages):
         self.logger.info("Processing %s alerts", str(len(messages)))
-        alerts_with_prv_candidates = list(
-            filter(
-                lambda alert: alert.get("extra_fields", {}).get("prv_candidates")
-                is not None,
-                messages,
-            )
-        )
         prv_detections, non_detections = process_prv_candidates(
-            self.prv_candidates_processor, alerts_with_prv_candidates
+            self.prv_candidates_processor, messages
         )
 
         return messages, prv_detections, non_detections
