+++ tests/integration/test_step.py
@@ -8,24 +8,46 @@ def test_step_initialization(kafka_service, env_variables):
     assert isinstance(step, PrvCandidatesStep)
 
 
-def test_step_start(kafka_service, env_variables):
+def test_result_has_everything(
+    kafka_service, env_variables, kafka_consumer: KafkaConsumer
+):
     from scripts.run_step import step
 
     step.start()
+    for message in kafka_consumer.consume():
+        assert_result_has_non_detections(message)
+        assert_result_has_prv_detections(message)
+        assert_result_has_alert(message)
+        kafka_consumer.commit()
 
-    consumer = KafkaConsumer(
-        {
-            "PARAMS": {
-                "bootstrap.servers": "localhost:9092",
-                "group.id": "test_step_start",
-                "auto.offset.reset": "beginning",
-                "enable.partition.eof": True,
-            },
-            "TOPICS": ["prv-candidates"],
-        }
-    )
 
-    for message in consumer.consume():
-        assert message["new_alert"] is not None
+def assert_result_has_prv_detections(message):
+    assert message["prv_detections"] is not None
+    if message["new_alert"]["tid"].lower() == "atlas":
+        assert len(message["prv_detections"]) == 0
+    else:
         assert len(message["prv_detections"]) == 2
+
+
+def assert_result_has_non_detections(message):
+    assert message["non_detections"] is not None
+    if message["new_alert"]["tid"].lower() == "atlas":
+        assert len(message["non_detections"]) == 0
+    else:
         assert len(message["non_detections"]) == 2
+
+
+def assert_result_has_alert(message):
+    assert message["new_alert"] is not None
+
+
+def test_scribe_has_non_detections():
+    pass
+
+
+def test_works_with_ztf_messages():
+    pass
+
+
+def test_works_with_atlas_message():
+    pass
