+++ tests/shared/prv_candidate_schema.avsc
@@ -1,0 +1,66 @@
+{
+	"namespace": "ztf.alert",
+	"name": "prv_candidate",
+	"doc": "avro alert schema",
+	"version": "3.3",
+	"type": "record",
+	"fields": [
+		{"name": "jd", "type": "double", "doc": "Observation Julian date at start of exposure [days]"},
+		{"name": "fid", "type": "int", "doc": "Filter ID (1=g; 2=R; 3=i)"},
+		{"name": "pid", "type": "long", "doc": "Processing ID for image"},
+		{"name": "diffmaglim", "type": ["float", "null"], "default": null, "doc": "Expected 5-sigma mag limit in difference image based on global noise estimate [mag]"},
+		{"name": "pdiffimfilename", "type": ["string", "null"], "default": null, "doc": "filename of positive (sci minus ref) difference image"},
+		{"name": "programpi", "type": ["string", "null"], "default": null, "doc": "Principal investigator attached to program ID"},
+		{"name": "programid", "type": "int", "doc": "Program ID: encodes either public, collab, or caltech mode"},
+		{"name": "candid", "type": ["long", "null"], "doc": "Candidate ID from operations DB"},
+		{"name": "isdiffpos", "type": ["string", "null"], "doc": "t or 1 => candidate is from positive (sci minus ref) subtraction; f or 0 => candidate is from negative (ref minus sci) subtraction"},
+		{"name": "tblid", "type": ["long", "null"], "default": null, "doc": "Internal pipeline table extraction ID"},
+		{"name": "nid", "type": ["int", "null"], "default": null, "doc": "Night ID"},
+		{"name": "rcid", "type": ["int", "null"], "default": null, "doc": "Readout channel ID [00 .. 63]"},
+		{"name": "field", "type": ["int", "null"], "default": null, "doc": "ZTF field ID"},
+		{"name": "xpos", "type": ["float", "null"], "default": null, "doc": "x-image position of candidate [pixels]"},
+		{"name": "ypos", "type": ["float", "null"], "default": null, "doc": "y-image position of candidate [pixels]"},
+		{"name": "ra", "type": ["double", "null"], "doc": "Right Ascension of candidate; J2000 [deg]"},
+		{"name": "dec", "type": ["double", "null"], "doc": "Declination of candidate; J2000 [deg]"},
+		{"name": "magpsf", "type": ["float", "null"], "doc": "Magnitude from PSF-fit photometry [mag]"},
+		{"name": "sigmapsf", "type": ["float", "null"], "doc": "1-sigma uncertainty in magpsf [mag]"},
+		{"name": "chipsf", "type": ["float", "null"], "default": null, "doc": "Reduced chi-square for PSF-fit"},
+		{"name": "magap", "type": ["float", "null"], "default": null, "doc": "Aperture mag using 14 pixel diameter aperture [mag]"},
+		{"name": "sigmagap", "type": ["float", "null"], "default": null, "doc": "1-sigma uncertainty in magap [mag]"},
+		{"name": "distnr", "type": ["float", "null"], "default": null, "doc": "distance to nearest source in reference image PSF-catalog [pixels]"},
+		{"name": "magnr", "type": ["float", "null"], "default": null, "doc": "magnitude of nearest source in reference image PSF-catalog [mag]"},
+		{"name": "sigmagnr", "type": ["float", "null"], "default": null, "doc": "1-sigma uncertainty in magnr [mag]"},
+		{"name": "chinr", "type": ["float", "null"], "default": null, "doc": "DAOPhot chi parameter of nearest source in reference image PSF-catalog"},
+		{"name": "sharpnr", "type": ["float", "null"], "default": null, "doc": "DAOPhot sharp parameter of nearest source in reference image PSF-catalog"},
+		{"name": "sky", "type": ["float", "null"], "default": null, "doc": "Local sky background estimate [DN]"},
+		{"name": "magdiff", "type": ["float", "null"], "default": null, "doc": "Difference: magap - magpsf [mag]"},
+		{"name": "fwhm", "type": ["float", "null"], "default": null, "doc": "Full Width Half Max assuming a Gaussian core, from SExtractor [pixels]"},
+		{"name": "classtar", "type": ["float", "null"], "default": null, "doc": "Star/Galaxy classification score from SExtractor"},
+		{"name": "mindtoedge", "type": ["float", "null"], "default": null, "doc": "Distance to nearest edge in image [pixels]"},
+		{"name": "magfromlim", "type": ["float", "null"], "default": null, "doc": "Difference: diffmaglim - magap [mag]"},
+		{"name": "seeratio", "type": ["float", "null"], "default": null, "doc": "Ratio: difffwhm / fwhm"},
+		{"name": "aimage", "type": ["float", "null"], "default": null, "doc": "Windowed profile RMS afloat major axis from SExtractor [pixels]"},
+		{"name": "bimage", "type": ["float", "null"], "default": null, "doc": "Windowed profile RMS afloat minor axis from SExtractor [pixels]"},
+		{"name": "aimagerat", "type": ["float", "null"], "default": null, "doc": "Ratio: aimage / fwhm"},
+		{"name": "bimagerat", "type": ["float", "null"], "default": null, "doc": "Ratio: bimage / fwhm"},
+		{"name": "elong", "type": ["float", "null"], "default": null, "doc": "Ratio: aimage / bimage"},
+		{"name": "nneg", "type": ["int", "null"], "default": null, "doc": "number of negative pixels in a 5 x 5 pixel stamp"},
+		{"name": "nbad", "type": ["int", "null"], "default": null, "doc": "number of prior-tagged bad pixels in a 5 x 5 pixel stamp"},
+		{"name": "rb", "type": ["float", "null"], "default": null, "doc": "RealBogus quality score; range is 0 to 1 where closer to 1 is more reliable"},
+		{"name": "ssdistnr", "type": ["float", "null"], "default": null, "doc": "distance to nearest known solar system object if exists within 30 arcsec [arcsec]"},
+		{"name": "ssmagnr", "type": ["float", "null"], "default": null, "doc": "magnitude of nearest known solar system object if exists within 30 arcsec (usually V-band from MPC archive) [mag]"},
+		{"name": "ssnamenr", "type": ["string", "null"], "default": null, "doc": "name of nearest known solar system object if exists within 30 arcsec (from MPC archive)"},
+		{"name": "sumrat", "type": ["float", "null"], "default": null, "doc": "Ratio: sum(pixels) / sum(|pixels|) in a 5 x 5 pixel stamp where stamp is first median-filtered to mitigate outliers"},
+		{"name": "magapbig", "type": ["float", "null"], "default": null, "doc": "Aperture mag using 18 pixel diameter aperture [mag]"},
+		{"name": "sigmagapbig", "type": ["float", "null"], "default": null, "doc": "1-sigma uncertainty in magapbig [mag]"},
+		{"name": "ranr", "type": ["double", "null"], "doc": "Right Ascension of nearest source in reference image PSF-catalog; J2000 [deg]"},
+		{"name": "decnr", "type": ["double", "null"], "doc": "Declination of nearest source in reference image PSF-catalog; J2000 [deg]"},
+		{"name": "scorr", "type": ["double", "null"], "default": null, "doc": "Peak-pixel signal-to-noise ratio in point source matched-filtered detection image"},
+		{"name": "magzpsci", "type": ["float", "null"], "default": null, "doc": "Magnitude zero point for photometry estimates [mag]"},
+		{"name": "magzpsciunc", "type": ["float", "null"], "default": null, "doc": "Magnitude zero point uncertainty (in magzpsci) [mag]"},
+		{"name": "magzpscirms", "type": ["float", "null"], "default": null, "doc": "RMS (deviation from average) in all differences between instrumental photometry and matched photometric calibrators from science image processing [mag]"},
+		{"name": "clrcoeff", "type": ["float", "null"], "default": null, "doc": "Color coefficient from linear fit from photometric calibration of science image"},
+		{"name": "clrcounc", "type": ["float", "null"], "default": null, "doc": "Color coefficient uncertainty from linear fit (corresponding to clrcoeff)"},
+		{"name": "rbversion", "type": "string", "doc": "version of RealBogus model/classifier used to assign rb quality score"}
+			]
+}
